Refactor markdown parsing
