+++ utils/markdown_utils.js
@@ -1,0 +1,47 @@
+function parseFrontMatter(text = '') {
+  if (!text.startsWith('---')) return { meta: {}, body: text };
+  const end = text.indexOf('\n---', 3);
+  if (end < 0) return { meta: {}, body: text };
+  const header = text.slice(3, end).trim();
+  const body = text.slice(end + 4);
+  const meta = {};
+  header.split(/\r?\n/).forEach(line => {
+    const parts = line.split(':');
+    const key = parts.shift().trim();
+    if (key) meta[key] = parts.join(':').trim();
+  });
+  return { meta, body };
+}
+
+function parseAutoIndex(text = '') {
+  const { meta } = parseFrontMatter(text);
+  if (!meta.files) {
+    const lines = text.split(/\r?\n/);
+    const idx = lines.findIndex(l => /^files:\s*/i.test(l.trim()));
+    if (idx >= 0) {
+      const arr = [];
+      for (let i = idx + 1; i < lines.length; i++) {
+        const ln = lines[i];
+        if (/^\s*-\s+/.test(ln)) {
+          arr.push(ln.replace(/^\s*-\s+/, '').trim());
+        } else if (/^[\w_-]+:\s*/.test(ln.trim())) {
+          break;
+        }
+      }
+      if (arr.length) meta.files = arr;
+    }
+  } else if (typeof meta.files === 'string') {
+    meta.files = meta.files
+      .replace(/^\[/, '')
+      .replace(/\]$/, '')
+      .split(/,|\r?\n/)
+      .map(t => t.replace(/^\s*-\s*/, '').trim())
+      .filter(Boolean);
+  }
+  return meta;
+}
+
+module.exports = {
+  parseFrontMatter,
+  parseAutoIndex,
+};
